Created and implemented PlayerFacade class.
Crteated PlayerFacade which implements IWallet, IInventory and IEquipment interfaces.
This was, not only for the sake of optimization by reducing the amount of MonoBehaviour attached components, but also to manage the communication between the Equipment system and the Inventory system.
Lasttly, I added assembly definitions for every single Script/Game/Gameplay/ folder.